added users role and staff management

+++ app/Http/Controllers/CustomerController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Customer;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class CustomerController extends Controller
 {
@@ -137,6 +138,10 @@ class CustomerController extends Controller
      */
     public function destroy(Customer $customer)
     {
+        if (Auth::user()->role !== 'admin') {
+            abort(403);
+        }
+
         $customer->delete();
 
         return redirect()
